finish InMemory with errors

+++ external-properties/src/main/scala/cn/pandadb/externalprops/InMemoryPropertyNodeStore.scala
@@ -25,27 +25,28 @@ object InMemoryPropertyNodeStore extends CustomPropertyNodeStore {
   def filterNodes(expr: NFPredicate): Iterable[NodeWithProperties] = {
     expr match {
       case NFGreaterThan(fieldName: String, value: AnyValue) =>
-        nodes.values.filter(x => x.field(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() >
+        nodes.values.filter(x => x.mutable().props.get(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() >
           value.asInstanceOf[NumberValue].doubleValue()).getOrElse(false))
 
-
-      case NFLessThan(fieldName: String, value: AnyValue) =>
+      /*case NFLessThan(fieldName: String, value: AnyValue) =>
         nodes.values.filter(x => x.field(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() <
-          value.asInstanceOf[NumberValue].doubleValue()).getOrElse(false))
-
+          value.asInstanceOf[NumberValue].doubleValue()).getOrElse(false))*/
+      case NFLessThan(fieldName: String, value: AnyValue) =>
+     nodes.values.filter(x => x.mutable().props.get(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() <
+       value.asInstanceOf[NumberValue].doubleValue()).getOrElse(false))
 
       case NFLessThanOrEqual(fieldName: String, value: AnyValue) =>
-        nodes.values.filter(x => x.field(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() <=
+        nodes.values.filter(x => x.mutable().props.get(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() <=
           value.asInstanceOf[NumberValue].doubleValue()).getOrElse(false))
 
 
       case NFGreaterThanOrEqual(fieldName: String, value: AnyValue) =>
-        nodes.values.filter(x => x.field(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() >=
+        nodes.values.filter(x => x.mutable().props.get(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() >=
           value.asInstanceOf[NumberValue].doubleValue()).getOrElse(false))
 
 
       case NFEquals(fieldName: String, value: AnyValue) =>
-        nodes.values.filter(x => x.field(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() ==
+        nodes.values.filter(x => x.mutable().props.get(fieldName).map(_.asInstanceOf[NumberValue].doubleValue() ==
           value.asInstanceOf[NumberValue].doubleValue()).getOrElse(false))
 
     }
@@ -75,11 +76,11 @@ object InMemoryPropertyNodeStore extends CustomPropertyNodeStore {
       }
       if (addedLabels != null && addedLabels.size>0) {
         nodes(nodeId).labels ++= addedLabels
-        nodes(nodeId).labels = nodes(nodeId).labels.toSet
+       // nodes(nodeId).labels = nodes(nodeId).labels.toSet
       }
       if (removedLabels != null && removedLabels.size>0) {
         val tmpLabels = nodes(nodeId).labels.toSet
-        nodes(nodeId).labels = tmpLabels -- removedLabels
+       // nodes(nodeId).labels = tmpLabels -- removedLabels
       }
 
   }
@@ -108,22 +109,12 @@ object InMemoryPropertyNodeStore extends CustomPropertyNodeStore {
   }
 
   override def beginWriteTransaction(): PropertyWriteTransaction = {
-    new BufferedExternalPropertyWriteTransaction() {
-      override def commitPerformer(): GroupedOpVisitor = {
-
-        new InMemoryGroupedOpVisitor(true, nodes)
-
-      }
-
-      override def rollbackPerformer(): GroupedOpVisitor = {
-        new InMemoryGroupedOpVisitor(false, nodes)
-      }
-    }
+    new BufferedExternalPropertyWriteTransaction(this, new InMemoryGroupedOpVisitor(true, nodes), new InMemoryGroupedOpVisitor(false, nodes))
   }
 }
 class InMemoryGroupedOpVisitor(isCommit: Boolean, nodes: mutable.Map[Long, NodeWithProperties]) extends GroupedOpVisitor {
 
-
+  var oldState = mutable.Map[Long, MutableNodeWithProperties]();
 
 
   override def start(ops: GroupedOps): Unit = {
@@ -134,22 +125,42 @@ class InMemoryGroupedOpVisitor(isCommit: Boolean, nodes: mutable.Map[Long, NodeW
   }
 
   override def visitAddNode(nodeId: Long, props: Map[String, Value], labels: Array[String]): Unit = {
+    if (isCommit) InMemoryPropertyNodeStore.addNodes(Iterable(NodeWithProperties(nodeId, props, labels)))
+    else {
+      InMemoryPropertyNodeStore.deleteNodes(Iterable(nodeId))
+    }
 
-      InMemoryPropertyNodeStore.addNodes(Iterable(NodeWithProperties(nodeId, props, labels)))
 
 
   }
 
   override def visitDeleteNode(nodeId: Long): Unit = {
-    InMemoryPropertyNodeStore.deleteNodes(Iterable(nodeId))
+    if (isCommit) InMemoryPropertyNodeStore.deleteNodes(Iterable(nodeId))
+    else {
+
+      val oldNode = oldState.get(nodeId).head
+      InMemoryPropertyNodeStore.addNodes(Iterable(NodeWithProperties(nodeId, oldNode.props.toMap, oldNode.labels)))
+
+    }
   }
 
   override def visitUpdateNode(nodeId: Long, addedProps: Map[String, Value],
                                updateProps: Map[String, Value], removeProps: Array[String],
                                addedLabels: Array[String], removedLabels: Array[String]): Unit = {
-
-    InMemoryPropertyNodeStore.updateNodes(nodeId: Long, addedProps: Map[String, Value],
+      if (isCommit) InMemoryPropertyNodeStore.updateNodes(nodeId: Long, addedProps: Map[String, Value],
       updateProps: Map[String, Value], removeProps: Array[String],
       addedLabels: Array[String], removedLabels: Array[String])
+      else {
+
+        val oldNode = oldState.get(nodeId).head
+        InMemoryPropertyNodeStore.addNodes(Iterable(NodeWithProperties(nodeId, oldNode.props.toMap, oldNode.labels)))
+
+      }
+  }
+
+  override def setOldState(oldState: mutable.Map[Long, MutableNodeWithProperties]): Unit = {
+
+    this.oldState = oldState
+
   }
 }
